inheritance on user store for requested reports
